requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-filter==2.1.0
django-heroku==0.3.1
djangorestframework==3.9.1
gunicorn==19.9.0
Markdown==3.0.1
numpy==1.16.1
pandas==0.24.1
psycopg2==2.7.7
python-dateutil==2.8.0
pytz==2018.9
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
whitenoise==4.1.2
